Always close channel passed to cache.WaitForSync()
This effectively works around a bug in
k8s.io/apimachinery/pkg/util/wait/wait.go:PollUntil, which will leak
active goroutines (spawned in poller()) if the channel is not stopped,
even after the wait condition is satisfied.

This bug will be pursued upstream and eventually the fix here can be
removed.

Fixes #204

Signed-off-by: Oren Shomron <shomron@gmail.com>