Create new context when shutting app down
There is a potential race that the app tries to use a closed context
when it shutting down. In the tests, this looks like a unexpected
error from ioutil.ReadAll. Fixes the issue by creating a new
shutdown context. I'm 99% sure this is the cause of this issue,
and I verified it by running 1000 iterations of the test with no
errors.

#403