Prefix custom column if matches an existing column
In custom resources definitions, there can be additional custom
columns. By default, the custom resource definition list printer
wants to pring Name, Labels, Age. If the CRD has an Age
column, it would overlap the existing Age column, and you'd
get a repeat. This change prefixes custom columns with "Resource"
and displays them.

This might not be the best solution. Perhaps custom columns
should be identified in a different way.

The PR also creates a new string util package and consolidates
the string slice contains function that existed in a couple places
in the app.