YAML Viewer & FlexLayout Title UX (#650)
We introduce trackBy in this commit. trackBy is a functionality that we should strive
to use whenever we use ngFor. It allows us to specify a function that Angular needs to
identify which item in a list of objects needs be rerendered when updating the DOM.
Angular by default uses object references, but this won't work when the list is
a new instance with new object references.

Lint is not 100% passing in this commit but we fix a few of the warnings here.