use a cached restmap instead of creating one on each cluster call
Previously, sugarloaf was creating a new restmapper for each cluster call. This regression included a ~3 second penalty for *each* get/list/watch due to having to find the current version of the api resource list. Replace the restmapper code with a `restmapper.DeferredDiscoveryRESTMapper` which will periodically update the api resource in the background.