Update README with launch expectation detail
Good to set expectations, the first few times I ran octant it didn't open a browser, or opened one in the background. I looked to see which port it would launch on but couldn't find it in the docs... so after getting it to work by trying a few more times, I realized it launches on a different port each time.  Just thought it would help someone else get started by setting the expecation a little =)