Check-in the generated mocks used for unit testing
Using Octant as a module in another Go project (e.g. when writing a
plugin) is currently broken. "go mod tidy" fails because it needs to
record test dependencies in the go.mod file, which is not possible for
Octant because test files refer to the generated mock packages ("fake"
packages) which are not checked-in. This change solves the issue by
checking-in the "fake" files generated by "go run build.go
generate". The change includes a CI check to ensure that mocks are
always up-to-date.

Fixes #427

Signed-off-by: Antonin Bas <abas@vmware.com>