crd printer should support multiple versions
In the printers, items are generated and then printed later in a func. This func is defined in a for loop and being referred to by value and not by reference, so each version ended up being the last version.

This change uses in the range index to ensure the proper version is being shown.

Noted in: https://github.com/vmware-tanzu/octant/issues/483#issuecomment-569144151

Signed-off-by: bryanl <bryanliles@gmail.com>