Fix logs in multi-context environments
Logging was broken for multi-cluster/multi-context environments.
Everything else in the dashboard loaded fine, except the logs.

It looks like passing the ClusterClient into the log handler ends up
caching a pointer to the original cluster. Once the user has switched
contexts from the UI, the underlying ClusterClient object is swapped
out on the dashConfig object. However, the pointer is cached for the
log handler and it still tries to read from the old cluster.

The fix seems to be passing the dashConfig pointer to the log
handler, and fetching the client on each call.

Tested against our 5 EKS clusters, and observed that logs were now
working after switching contexts.

Signed-off-by: Jack Wink <57678801+mothershipper@users.noreply.github.com>