Extract creation of frontend/backend handlers
This change extracts the creation of Octant's frontend and backend
handlers from pkg/dash. It also introduces a new namespace,
github.com/vmware-octant/pkg/octant. This package will eventually become
the home for Octant's top level domain objects.

A new options construct is introducted in `pkg/octant`. It's objective
is to persistent a consistent set of options that can be used to
configure Octant's behavior. In the initial version, fields for frontend
and backend HTTP handlers have been added.

This change is transitionary as its purpose to start the process of
detangling `pkg/dash`.

Signed-off-by: bryanl <bryanliles@gmail.com>