Apply YAML UI
Adds a link in the header to an empty YAML editor that uploads resources
to the cluster and defines an action `action.octant.dev/apply` which
parses the YAML and creates or updates each resources on the cluster.

Resources are processed and applied in the order they are defined.
New resources are created and existing resources are patched with the
changes applied server side, which requires k8s 1.16+.

Errors will stop the processing of new resources. Successfully applied
resources will be reported alongside any errors. When multiple resources
are applied, a summary is reported. Errors are lightly processed and may
not be user friendly.

Namespaced resources that do not define a namespace are defaulted to
the current namespace.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>