Switch Mutating/ValidatingWebhookConfiguration version to v1beta1
Cluster before Kubernetes 1.16 don't support admissionregistration.k8s.io/v1.
The v1beta1 version was deprecated in k8s 1.16, but is still available
and semantically equivalent.

This change also fixes a bug in the ObjectStore where it might return
objects for a different version than requested. Instead of the requested
version, the server's default version for the resource is used. With
this change, it is possible that we will have multiple informers for the
same resource at different versions. While inefficient the conversion
logic is on the server, where it belongs.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>