Fix default theme on load
Octant should use the light theme by default, however, it was using dark
as a quirk of how the starting theme was set. This led to the editor
using a different theme than Octant.

Now the light theme is used consistently by default. A previously
toggled theme is preserved.

This change reduces the number of places where the theme state is
stored, and synchronized. In particular, we now avoid unnecessary
lookups from local storage, which are quite slow due to being
synchronous disk i/o.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>