Display a 404-like page if a resource is not found (#1252)
Previously, Octant would redirect to a parent path segment when a
resource is not found. This could be because the path is not mapped in
Octant, the particular namespace or resource is not found, or in the
case of CRDs, the CRD is not defined. This behavior means that the URL a
user navigated to may not be the URL that is rendered.

Now, the URL is preserved and an error page is displayed instead. The
user is present with a small number of options including going to the
parent resource, or waiting. If the user remains on the page and the
resource becomes valid, the new resource will replace the error page.
This behavior is particularly useful when watching a single resource
that comes and goes.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>