Add a ButtonGroup to the Table
Tables support grid actions that perform actions on a given row of the
table, but did not have a way to take an action on the whole table. Now
an optional button group is rendered to the right of the title.

For example, a button could be used to create a new item in the table.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>