Add support for selectors in js dashboard client list
When the js key is converted to an octant key, the selector information was elided. This change adds support for key `selectors` and `labelSelectors`.

In support of this change, a new Raw function was added to Payload.

Also, the RequireErrorOrNot test utility was updated to accept an optional list of functions to run when an error was not encountered. With this change, you can write simpler error checking code:

```go
// old
got, err := someCode()
if tt.wantErr {
  require.Error(t, err)
  return
}
require.NoError(t, err)
runCheck(got)

// new
got, err := someCode()
testutil.RequireErrorOrNot(t, tt.wantErr, err, func() {
  runCheck(got)
})
```

Signed-off-by: bryanl <bryanliles@gmail.com>