Move websocket service to new module
This change moves the websocket service to a new module to make it
easier to test and work with. In the process of moving the service, a
new token, `WindowToken` has been created that allows the ability to
provide custom values for `window`.

This change als updates the typescript config to import the `dom` lib.
This hint provides editor access to dom provided globals.

Signed-off-by: bryanl <bryanliles@gmail.com>