add startup integration tests
Also externalize NewRunner's dependence on net.Listener and inject an in-memory
listener for the integration tests -- this should make them a bit faster,
stabler (no dependence on the operating system to provision a TCP socket) and
safe to run in parallel.

Really just covering the happy paths here. Ideally some function signatures
can get cleaned up and potentially an isolate-able 'startup component' could
be extracted to get more focused unit-test feedback.

The 'startup without a kubeconfig' flow was a bit more delicate -- after sending
the kubeconfig over the streaming API, there is a brief period of downtime
while the API reloads and the test had to wait for this.

There is one possible edge case where the "clean shutdown before receiving
kubeconfig" scenario will erroneously pass -- this is in the case that the
machine running the tests contains a valid kubeconfig at the default location,
and the test takes over a second to run.

In this case the LoadingManager's behaviour of loading a kubeconfig from the
default location (1 second after starting) would kick in, and the API would
successfully be initialized -- therefore we would not be exercising the
"shutdown before loading kubeconfig" scenario we claim to exercise.

We tried refactoring various packages to the point that we could inject a fake
filesystem into the LoadingManager (to force it to never find a valid kubeconfig
at the default location) but it was fairly intrusive, and it seems exceedingly
rare that our new test should take anywhere near 1 second to run.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>
Co-authored-by: Dimitry Besson <dbesson@vmware.com>
Co-authored-by: James Thomson <jamesth@vmware.com>
Co-authored-by: Vikram Yadav <yvikram@vmware.com>