encapsulate ocontext.WithKubeConfigCh
allowing users to simply call `NewRunner` without having to remember to populate
the passed context with a kubeconfig channel.

We achieve this by having the Runner keep track of the context it was created
with. While we were at it, we noticed that there was no need to have a Logger
parameter for Start -- in fact, in the long run we may be able to keep removing
parameters from Start until only the startup/shutdown channels remain.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>
Co-authored-by: James Thomson <jamesth@vmware.com>