Fix error when closing browser tab
There was some incorrect error handling in my last commit. When fixing
that I noticed that I slightly changed the semantics of the websocket
read pump when I moved unmarshaling to be a client.Recieve
responsibility. So I introduced a new concept that a stream client can
return a fatal error, and read pump can react to that appropriately.
This paves the way for read pump to become a function owned by the
streaming client manager instead of a method on websocket client.

Signed-off-by: Michael Stergianis <mstergianis@vmware.com>